requirements.txt
torch>=1.13.1
torchvision>=0.14.1
timm>=0.6.12
albumentations>=1.3.0
scikit-image>=0.19.3
scikit-learn>=1.1.3
PyYAML>=6.0
opencv-python>=4.7.0
matplotlib>=3.6.2
numpy>=1.22.0